backend/requirements.txt
asgiref==3.8.1
dj-database-url==2.1.0
Django==4.2.11
django-cors-headers==4.3.1
djangorestframework==3.15.2
gunicorn==22.0.0
psycopg==3.2.4
python-dotenv==1.0.1
whitenoise==6.6.0
